requirements.txt
boto3==1.26.137
python-dotenv==1.0.0
Pillow==10.0.0
pytest==7.4.0
pytest-cov==4.1.0
moto==4.1.12
localstack-client==2.0
requests==2.31.0
